requirements.txt
django==3.2.15
django-debug-toolbar==3.2.1
pillow==11.1.0
environs[django]==9.3.2
django-phonenumber-field==8.0.0
phonenumbers==8.13.54
djangorestframework==3.13.0
